refactor: GameController의 startMove 에서 결과반환과 움직임 판단(루프문) 분리
- 그에 따른 move의 valid 판단 뿐만 아니라 사용자가 선택한 다리 정보를 저장하는 로직도 move 로직 내부로 이동
	-> 그로 인해 BridgeGame과 GameService 간의 유사도가 높아지는 것을 확인하여 BridgeGame 클래스 하나로 통일, 그에 따른 테스트 코드도 수정
